update impl to use StandardCharacter
