+++ lua/wikis/commons/Widget/CharacterTable.lua
@@ -9,8 +9,9 @@
 local Array = require('Module:Array')
 local Class = require('Module:Class')
 local Lua = require('Module:Lua')
-local Operator = require('Module:Operator')
 local String = require('Module:StringUtils')
+
+local Character = Lua.import('Module:Character')
 
 local Widget = Lua.import('Module:Widget')
 local HtmlWidgets = Lua.import('Module:Widget/Html/All')
@@ -24,23 +25,17 @@ local CharacterTable = Class.new(Widget)
 
 ---@return Widget
 function CharacterTable:render()
-	local conditions = '[[type::character]]'
 	local extraConditions = self.props.extraConditions
 	if String.isNotEmpty(extraConditions) then
 		conditions = conditions .. ' AND (' .. extraConditions .. ')'
 	end
-	local characterNames = Array.map(mw.ext.LiquipediaDB.lpdb('datapoint', {
-		conditions = conditions,
-		query = 'name',
-		order = 'name asc',
-		limit = 1000,
-	}), Operator.property('name'))
+	local characters = Character.getAllCharacters()
 	return Div{
 		css = {
 			['text-align'] = 'center'
 		},
-		children = Array.interleave(Array.map(characterNames, function (characterName)
-			return Entry{name = characterName, alias = self.props.alias}
+		children = Array.interleave(Array.map(characters, function (character)
+			return Entry(character)
 		end), '\n')
 	}
 end
