+++ lua/wikis/commons/Widget/CharacterTable/Entry/Custom.lua
@@ -1,0 +1,11 @@
+---
+-- @Liquipedia
+-- wiki=commons
+-- page=Module:Widget/CharacterTable/Entry/Custom
+--
+-- Please see https://github.com/Liquipedia/Lua-Modules to contribute
+--
+
+local Lua = require('Module:Lua')
+
+return Lua.import('Module:Widget/CharacterTable/Entry')
